Module sodiumoxide::crypto

Cryptographic functions

Modules

aead
Authenticated Encryption with Additional Data This operation:

auth
Secret-key authentication

box_
Public-key authenticated encryption

generichash
GenericHash.

hash
Hashing

kdf
Key derivation

kx
Key exchange

onetimeauth
Secret-key One-time authentication

pwhash
Password Hashing

scalarmult
Scalar multiplication

sealedbox
Sealed Boxes

secretbox
Secret-key authenticated encryption

secretstream
Stream encryption/file encryption

shorthash
A lot of applications and programming language implementations have been recently found to be vulnerable to denial-of-service attacks when a hash function with weak security guarantees, like Murmurhash 3, was used to construct a hash table.

sign
Public-key signatures

stream
Secret-key encryption

verify
Constant-time comparison of fixed-size vecs