Function sodiumoxide::crypto::box_::curve25519xsalsa20poly1305::open_detached_precomputed

pub fn open_detached_precomputed(
    c: &mut [u8], 
    mac: &Tag, 
    n: &Nonce, 
    k: &PrecomputedKey
) -> Result<(), ()>

open_detached_precomputed() verifies and decrypts a ciphertext c using a precomputed key k and a nonce n. c is decrypted in place, so if this function is successful it will contain the plaintext. If the ciphertext fails verification, open_detached() returns Err(()), and the ciphertext is not modified.