sodiumoxide::crypto::generichash
Struct State

pub struct State { /* private fields */ }

State contains the state for multi-part (streaming) hash computations. This allows the caller to process a message as a sequence of multiple chunks.

Implementations

impl State

pub fn new(out_len: usize, key: Option<&[u8]>) -> Result<State, ()>

new constructs and initializes a new State with the given parameters.

out_len specifies the resulting hash size. Only values in the interval [DIGEST_MIN, DIGEST_MAX] are allowed.

key is an optional parameter, which when given, a custom key can be used for the computation of the hash. The size of the key must be in the interval [KEY_MIN, KEY_MAX].

pub fn update(&mut self, data: &[u8]) -> Result<(), ()>

update updates the State with data. update can be called multiple times in order to compute the hash from sequential chunks of the message.

pub fn finalize(self) -> Result<Digest, ()>

finalize finalizes the state and returns the digest value. finalize consumes the State so that it cannot be accidentally reused.