Module sodiumoxide::crypto::hash::sha256

SHA-256.

There has been considerable degradation of public confidence in the security conjectures for many hash functions, including SHA-256. However, for the moment, there do not appear to be alternatives that inspire satisfactory levels of confidence. One can hope that NIST’s SHA-3 competition will improve the situation.

Structs

Digest
Digest-structure

State
State contains the state for multi-part (streaming) hash computations. This allows the caller to process a message as a sequence of multiple chunks.

Constants

BLOCKBYTES
Block size of the hash function.

DIGESTBYTES
Number of bytes in a Digest.

Functions

hash
hash hashes a message m. It returns a hash h.