Struct sodiumoxide::crypto::hash::sha512::State

pub struct State(_);

State contains the state for multi-part (streaming) hash computations. This allows the caller to process a message as a sequence of multiple chunks.

Methods

impl State

pub fn new() -> Self

new constructs and initializes a new State.

pub fn update(&mut self, data: &[u8])

update updates the State with data. update can be called multiple times in order to compute the hash from sequential chunks of the message.

pub fn finalize(self) -> Digest

finalize finalizes the state and returns the digest value. finalize consumes the State so that it cannot be accidentally reused.