Module sodiumoxide::crypto::kdf::blake2b

blake2b is the current default key derivation scheme of libsodium.

Structs

Key
Key for key derivation.

Constants

BYTES_MAX
Maximum number of bytes in a subkey.

BYTES_MIN
Minimum number of bytes in a subkey.

CONTEXTBYTES
Number of bytes in a Context

KEYBYTES
Number of bytes in a Key.

Functions

derive_from_key
derive_from_key derives the subkey_id-th subkey from the master key key and the context ctx

gen_key
gen_key() randomly generates a key for key derivation.