sodiumoxide::crypto::kx::x25519blake2b
Function client_session_keys

pub fn client_session_keys(
    client_pk: &PublicKey,
    client_sk: &SecretKey,
    server_pk: &PublicKey,
) -> Result<(SessionKey, SessionKey), ()>

client_session_keys() computes a pair of shared keys (rx and tx) using the client’s public key client_pk, the client’s secret key client_sk and the server’s public key server_pk. If the server’s public key is acceptable, it returns the two shared keys, the first for rx and the second for tx. Otherwise, it returns None.