Module sodiumoxide::crypto::sealedbox::curve25519blake2bxsalsa20poly1305

A particular combination of Curve25519, Blake2B, Salsa20 and Poly1305.

Functions

open
The open() function decrypts the ciphertext c using the key pair (pk, sk) and returns the decrypted message.

seal
The seal() function encrypts a message m for a recipient whose public key is pk. It returns the ciphertext whose length is SEALBYTES + m.len().