Module sodiumoxide::crypto::secretstream::xchacha20poly1305::builder_pattern

Potential convenience builder pattern.

Adds a single message method to the Stream to create a MessageBuilder. Allows for chaining parameters, and defaults to using Tag::Message. for the tag.

Example:

use sodiumoxide::crypto::secretstream::builder_pattern as secretstream;
use self::secretstream::{Key, Tag};

let pt = b"my plaintext bytes";
let ad = b"some additional data";

let key = Key::new();
let (mut stream, header) = secretstream::init_push(&key)?;
let ct1 = stream.message(pt).push()?;
let ct2 = stream.message(pt).ad(ad).push()?;
let ct3 = stream.message(pt).tag_final().push()?;
// ... some time later ...
let mut stream2 = secretstream::init_pull(&header, &key)?;
let (rec1, tag1) = stream2.pull(&ct1, None)?;
let (rec2, tag2) = stream2.pull(&ct2, Some(ad))?;
let (rec3, tag3) = stream2.pull(&ct3, None)?;

assert_eq!(rec1, pt);
assert_eq!(tag1, Tag::Message);
assert_eq!(rec2, pt);
assert_eq!(tag2, Tag::Message);
assert_eq!(rec3, pt);
assert_eq!(tag3, Tag::Final);

Re-exports

pub use super::oop::*;

Structs

MessageBuilder
A builder for producing secret stream messages with the correct tags and authenticated data.