Module sodiumoxide::crypto::secretstream::xchacha20poly1305::low_level

Simple wrapper around lower level libsodium functionality, with minor attempts to make the bindings idiomatic Rust.

Provides basic functions, and a thin wrapper for State. Does not provide any type-checking for Push/Pull states.

Example:

use sodiumoxide::crypto::secretstream::low_level as secretstream;
use self::secretstream::{Key, Tag};

let pt = b"my plaintext bytes";
let key = Key::new();
let (mut state1, header) = secretstream::init_push(&key)?;
let ct = secretstream::push(pt, None, Tag::Message, &mut state1)?;
// ... some time later ...
let mut state2 = secretstream::init_pull(&header, &key)?;
let (rec, tag) = secretstream::pull(&ct, None, &mut state2)?;

assert_eq!(rec, pt);
assert_eq!(tag, Tag::Message);

Note that calling push or pull with the wrong type of state is only a runtime error.

use sodiumoxide::crypto::secretstream::low_level as secretstream;
use self::secretstream::{Key, Tag};

let pt = b"my plaintext bytes";
let key = Key::new();
let (mut state, header) = secretstream::init_push(&key)?;
assert!(secretstream::pull(&[], None, &mut state).is_err());

Re-exports

pub use super::*;

Structs

State
Wrapper for the secretstream state. Used when pushing or pulling messages onto/from the stream.

Functions

init_pull
Initialize the secretstream state for pulling messages.

init_push
Initialize the secretstream state for pushing messages.

pull
Pull an encrypted message from the stream, verify and decrypt it.

pull_buf
Pull an encrypted message from the stream, verify and decrypt it.

push
All data (including optional fields) is authenticated. Encrypts a message m and its tag. Optionally includes additional data ad, which is not encrypted.

push_buf
All data (including optional fields) is authenticated. Encrypts a message m and its tag. Optionally includes additional data ad, which is not encrypted.

rekey⚠
Explicit rekeying. This updates the internal state of the State, and should only be called in a synchronized manner with how the corresponding State called it when encrypting the stream.