Module sodiumoxide::crypto::secretstream::xchacha20poly1305::oop

A slightly higher level abstraction using OOP methods

Instead of using the raw State, a new Stream object is provided, which contains the stream state, as well as tracking whether the stream has been finalized. Stream additionally has a type marker StreamMode to track whether it can be used for pushing or pulling. Note that the StreamMode trait cannot be implemented externally.

Additionally rotates the key on dropping the Stream to clear the old keys from memory.

Example:

use sodiumoxide::crypto::secretstream::oop as secretstream;
use self::secretstream::{Key, Tag};

let pt = b"my plaintext bytes";
let ad = b"some additional data";

let key = Key::new();
let (mut stream, header) = secretstream::init_push(&key)?;
let ct1 = stream.push(pt, None, Tag::Message)?;
let ct2 = stream.push(pt, Some(ad), Tag::Message)?;
let ct3 = stream.push(pt, None, Tag::Final)?;
// ... some time later ...
let mut stream2 = secretstream::init_pull(&header, &key)?;
let (rec1, tag1) = stream2.pull(&ct1, None)?;
let (rec2, tag2) = stream2.pull(&ct2, Some(ad))?;
let (rec3, tag3) = stream2.pull(&ct3, None)?;

assert_eq!(rec1, pt);
assert_eq!(tag1, Tag::Message);
assert_eq!(rec2, pt);
assert_eq!(tag2, Tag::Message);
assert_eq!(rec3, pt);
assert_eq!(tag3, Tag::Final);

Compared to low_level, trying to use the wrong stream operation is a compile-time error:

use sodiumoxide::crypto::secretstream::oop as secretstream;
use self::secretstream::{Key, Tag};

let pt = b"my plaintext bytes";
let key = Key::new();
let (mut stream, header) = secretstream::init_push(&key)?;
let _ = stream.pull(&[], None).unwrap();

Fails with error[E0599]: no method named `pull` found for type `Stream<Push>` in the current scope

Re-exports

pub use super::*;

Structs

Stream
Stream contains the state for multi-part (streaming) computations. This allows the caller to process encryption of a sequence of multiple messages.

Enums

Pull

Push

Traits

StreamMode

Functions

init_pull
Initialize the secretstream for pulling messages.

init_push
Initialize the secretstream for pushing messages.