Module sodiumoxide::crypto::secretstream::xchacha20poly1305::with_traits

Alternative convenience functions around Iterator and Read/Write.

Current limitations: cannot

Examples

With iterators

Turns an iterator of &[u8] slies into an iterator of encrypted chunks Vec<u8> and vice-versa. An optional additional data can be specified for the entire iterator.

use sodiumoxide::crypto::secretstream::with_traits as secretstream;
use self::secretstream::{Key, Tag};

let pts = vec![vec![1u8], vec![2u8, 3u8], vec![4u8, 5u8, 6u8]];
let ad = b"some additional data";

let key = Key::new();
let (mut stream, header) = secretstream::init_push(&key)?;
let mut stream_iter = stream.wrap_iterator(&pts, Some(ad));
let cts = stream_iter.collect::<Result<Vec<Vec<u8>>, ()>>()?; // errors if any `push` fails
// ... some time later ...
let mut stream2 = secretstream::init_pull(&header, &key)?;
let mut stream2 = stream2.wrap_iterator(cts, Some(ad));
let rec = stream2.map(|res| res.map(|(rec, _t)| rec)).collect::<Result<Vec<Vec<u8>>, ()>>()?;
assert_eq!(pts, rec);

With Read/Write

Wraps a Write object with a Stream<Push>, so the encrypted stream is written directly into the inner Write.

Similarly, wraps a Read object with a Stream<Pull>, so that the decrypted stream can be read out directly.

A message size must be specified for both, so that the correct chunks can be parsed. Optional additional data can be specified for the entire stream.

use sodiumoxide::crypto::secretstream::with_traits as secretstream;
use self::secretstream::{Key, Tag};
use std::io::{Cursor, Read, Write};

let pt = b"my plaintext bytes";
let ad = b"some additional data";

let key = Key::new();
let (mut stream, header) = secretstream::init_push(&key)?;
let mut output = Cursor::new(Vec::new());
{
    let mut stream_w = stream.wrap_writer(&mut output, Some(ad), 16);
    stream_w.write_all(pt).unwrap();
}
// ... some time later ...
let mut stream2 = secretstream::init_pull(&header, &key)?;
let mut input = Cursor::new(output.get_ref().clone());
let mut stream_r = stream2.wrap_reader(&mut input, Some(ad), 16);
let mut rec = Vec::new();
stream_r.read_to_end(&mut rec).unwrap();
assert_eq!(pt, &rec[..]);

Re-exports

pub use super::oop::*;

Structs

SecretStreamIter

StreamIo