Module sodiumoxide::crypto::sign::edwards25519sha512batch

WARNING: This signature software is a prototype. It has been replaced by the final system Ed25519. It is only kept here for compatibility reasons.

Structs

PublicKey
PublicKey for signatures

SecretKey
SecretKey for signatures

Constants

PUBLICKEYBYTES
Number of bytes in a PublicKey.

SECRETKEYBYTES
Number of bytes in a SecretKey.

SIGNATUREBYTES
Number of bytes in a Signature.

Functions

gen_keypair
gen_keypair() randomly generates a secret key and a corresponding public key.

sign
sign() signs a message m using the signer's secret key sk. sign() returns the resulting signed message sm.

verify
verify() verifies the signature in sm using the signer's public key pk. verify() returns the message Ok(m). If the signature fails verification, verify() returns Err(()).