Struct sodiumoxide::crypto::stream::salsa208::Key

#[must_use]

pub struct Key(pub [u8; 32]);

Key for symmetric encryption

When a Key goes out of scope its contents will be zeroed out

Methods

impl Key

pub fn from_slice(bs: &[u8]) -> Option<Key>

from_slice() creates an object from a byte slice

This function will fail and return None if the length of the byte-slice isn't equal to the length of the object