Function sodiumoxide::utils::increment_le

pub fn increment_le(x: &mut [u8])

increment_le() treats x as an unsigned little-endian number and increments it in constant time.

WARNING: this method does not check for arithmetic overflow. When used for incrementing nonces it is the caller's responsibility to ensure that any given nonce value is used only once. If the caller does not do that the cryptographic primitives in sodiumoxide will not uphold any security guarantees (i.e. they may break)