Function sodiumoxide::crypto::pwhash::scryptsalsa208sha256::pwhash

pub fn pwhash(passwd: &[u8], OpsLimit: OpsLimit, MemLimit: MemLimit) -> Option<HashedPassword>

The pwhash() returns a HashedPassword which includes:

the result of a memory-hard, CPU-intensive hash function applied to the password passwd

the automatically generated salt used for the previous computation

the other parameters required to verify the password: opslimit and memlimit

OPSLIMIT_INTERACTIVE and MEMLIMIT_INTERACTIVE are safe baseline values to use for opslimit and memlimit.

The function returns Some(hashed_password) on success and None if it didn't complete successfully

Example

use sodiumoxide::crypto::pwhash::{pwhash, HashedPassword,
                                  OPSLIMIT_INTERACTIVE, MEMLIMIT_INTERACTIVE};
let passwd = "Correct Horse Battery Staple".as_bytes();
let pwh = pwhash(passwd, OPSLIMIT_INTERACTIVE, MEMLIMIT_INTERACTIVE).unwrap();
let pwh_bytes = &pwh[..];
//store pwh_bytes somewhere